Move the indexes get to the put function to minimises the i/o gap
Most likely will still need to 'fan in' via an SQS and possibly use a dynamodb table
