+++ fotos/common/invokePutIndex.ts
@@ -4,24 +4,25 @@ import {
 import { INVOCATION_EVENT } from "../lib/constants";
 import lambda from "../lib/lambda";
 import {
-  IIndex,
+  IIndexUpdate, IPutIndexRequest,
 } from "../types";
 
-export function getInvokeUpdateParams(body: IIndex, traceMeta): InvocationRequest {
+export function getInvokeUpdateParams(indexUpdate: IIndexUpdate, traceMeta): InvocationRequest {
+  const body: IPutIndexRequest = {
+    indexUpdate,
+    traceMeta,
+  };
   return {
     FunctionName: `${process.env.LAMBDA_PREFIX}indexesUpdate`,
     InvocationType: INVOCATION_EVENT,
     LogType: "Tail",
     Payload: JSON.stringify({
-      body: JSON.stringify({
-        index: body,
-        traceMeta,
-      }),
+      body: JSON.stringify(body),
     }),
   };
 }
 
-export default function invokePutIndex(body: IIndex, traceMeta) {
-  const params = getInvokeUpdateParams(body, traceMeta);
+export default function invokePutIndex(indexUpdate: IIndexUpdate, traceMeta) {
+  const params = getInvokeUpdateParams(indexUpdate, traceMeta);
   return lambda.invoke(params).promise();
 }
