+++ fotos/stream.ts
@@ -17,7 +17,7 @@ import invokePutIndex from "./common/invokePutIndex";
 import { safeLength } from "./create";
 import { getZeroCount } from "./indexes";
 import {
-  IIndex, IIndexDictionary, IIndexFields, ILoggerBaseParams,
+  IIndex, IIndexDictionary, IIndexFields, IIndexUpdate, ILoggerBaseParams,
 } from "./types";
 
 export function normaliseArrayFields(record: DynamoDBRecord): IIndexFields {
@@ -46,7 +46,7 @@ export function normaliseArrayFields(record: DynamoDBRecord): IIndexFields {
   return arrayFields;
 }
 
-export function parseIndexes(records: DynamoDBRecord[]): IIndex {
+export function getIndexUpdates(records: DynamoDBRecord[]): IIndexUpdate {
   return records.reduce((indexes, record) => {
     const updatedIndexes = { ...indexes };
     const arrayFields = normaliseArrayFields(record);
@@ -74,36 +74,9 @@ todo
 remove people from indexes, just handle tags here as
 the info on people is all in people.json inc number of images (faces)
 so then indexes just becomes tags - makes it scalable, add another index file
-for another taxoniomy if we adds one
+for another taxonomy if we adds one
 could also make tags have an array of images to make for more accuarte counts (and easier auditing)
 */
-
-export function updateCounts(existing: IIndex, newUpdates: IIndex) {
-  const updated: IIndex = {
-    people: {},
-    tags: {},
-  };
-  ["tags", "people"].forEach((key) => {
-    updated[key] = { ...existing[key] };
-    Object.keys(newUpdates[key]).forEach((item) => {
-      updated[key][item] = updated[key][item] ?
-        Math.max(0, updated[key][item] + newUpdates[key][item]) :
-        Math.max(0, newUpdates[key][item]);
-    });
-  });
-  return updated;
-}
-
-export function getUpdatedIndexes(existing: IIndex, newRecords: DynamoDBRecord[]): IIndex {
-  const updates: IIndex = parseIndexes(newRecords);
-  return updateCounts(existing, updates);
-}
-
-export function getModifiedIndexItems(existing: IIndexDictionary, updated: IIndexDictionary) {
-  return Object.keys(updated).filter((fieldKey: string) => {
-    return updated[fieldKey] !== existing[fieldKey];
-  });
-}
 
 export function getFirstRecord(records: DynamoDBRecord[]) {
   const firstRecord: DynamoDBRecord = records[0];
@@ -112,7 +85,7 @@ export function getFirstRecord(records: DynamoDBRecord[]) {
     ddbAttVals.unwrap(records[0].dynamodb!.OldImage);
 }
 
-export function getLogFields(records: DynamoDBRecord[], existingIndex?: IIndex, updatedIndexes?: IIndex) {
+export function getLogFields(records: DynamoDBRecord[], indexUpdates?: IIndexUpdate) {
   const firstRecord = getFirstRecord(records);
   return {
     ddbEventName: records[0].eventName,
@@ -132,18 +105,8 @@ export function getLogFields(records: DynamoDBRecord[], existingIndex?: IIndex, 
     imageUserIdentityId: firstRecord.userIdentityId,
     imageUsername: firstRecord.username,
     imageWidth: firstRecord.meta && firstRecord.meta.width,
-    indexesModifiedPeopleCount: existingIndex && updatedIndexes &&
-      getModifiedIndexItems(existingIndex.people, updatedIndexes.people).length,
-    indexesModifiedTagCount: existingIndex && updatedIndexes &&
-      getModifiedIndexItems(existingIndex.tags, updatedIndexes.tags).length,
-    indexesPeopleCount: existingIndex && Object.keys(existingIndex.people).length,
-    indexesTagCount: existingIndex && Object.keys(existingIndex.tags).length,
-    indexesUpdatedPeopleCount: updatedIndexes && Object.keys(updatedIndexes.people).length,
-    indexesUpdatedTagCount: updatedIndexes && Object.keys(updatedIndexes.tags).length,
-    indexesUpdatedZeroPeopleCount: updatedIndexes && getZeroCount(updatedIndexes.people),
-    indexesUpdatedZeroTagCount: updatedIndexes && getZeroCount(updatedIndexes.tags),
-    indexesZeroPeopleCount: existingIndex && getZeroCount(existingIndex.people),
-    indexesZeroTagCount: existingIndex && getZeroCount(existingIndex.tags),
+    indexesUpdatedPeopleCount: indexUpdates && Object.keys(indexUpdates.people).length,
+    indexesUpdatedTagCount: indexUpdates && Object.keys(indexUpdates.tags).length,
   };
 }
 
@@ -157,10 +120,9 @@ export async function indexRecords(event: DynamoDBStreamEvent, context: Context,
     traceId: uuid.v1(),
   };
   try {
-    const existingIndex: IIndex = await invokeGetIndex(getTraceMeta(loggerBaseParams));
-    const updatedIndexes: IIndex = getUpdatedIndexes(existingIndex, event.Records);
-    const response: InvocationResponse = await invokePutIndex(updatedIndexes, getTraceMeta(loggerBaseParams));
-    logger(context, loggerBaseParams, getLogFields(event.Records, existingIndex, updatedIndexes));
+    const indexUpdates: IIndex = getIndexUpdates(event.Records);
+    const response: InvocationResponse = await invokePutIndex(indexUpdates, getTraceMeta(loggerBaseParams));
+    logger(context, loggerBaseParams, getLogFields(event.Records, indexUpdates));
     return callback(null, success(response));
   } catch (err) {
     logger(context, loggerBaseParams, { err, ...getLogFields(event.Records) });
